refactor: one download handler for the single-quality platforms
